Switch to freespace seeking planner
Pick goals from a subset of all possible angles. Select according to the lowest cost, then the angle that requires the least rotation
This is only slightly more premeditated than the previous method of simply picking a nearby random angle to go in, but it saves a lot of time on the real robot because rejecting the random angle (once the local planner and the recovery actions bottom out) takes some time. Now we pick a direction we should be able to reach
Most of the raytracing code copied from costmap 2D header beacuse it was protected