Tweak costmap params
Drop inf observations. This stops the corners of the costmap from being filled with max observations
Use two obstacle layers, one for bump one for scan. I wasn't able to get the voxel layer to respect the z dimension of the scan and bump observations, so there was no way to prevent the scan data from clearing the bumps
Increase costmap size